+++ lib/shex/shape_expressions/node_constraint.ex
@@ -75,6 +75,8 @@ defmodule ShEx.NodeConstraint do
   end
 
   defimpl ShEx.Operator do
+    def children(_), do: []
+
     def triple_expression_label_and_operands(_), do: {nil, []}
   end
 end