+++ CHANGELOG.md
@@ -5,6 +5,17 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 [Keep a CHANGELOG](http://keepachangelog.com).
 
 
+## Unreleased
+
+### Added 
+
+- Checking unsatisfied references during schema parsing and creation 
+
+
+[Compare v0.1.0...HEAD](https://github.com/marcelotto/shex-ex/compare/v0.1.0...HEAD)
+
+
+
 ## v0.1.0 - 2019-07-15
 
 Initial release